# [BugFix] fix the bug of output profile to BE log